fix: moved registration tests to specfic files.
