+++ src/register/RegistrationPage.test.jsx
@@ -11,19 +11,16 @@ import { mockNavigate, BrowserRouter as Router } from 'react-router-dom';
 import renderer from 'react-test-renderer';
 import configureStore from 'redux-mock-store';
 
-import RegistrationFailureMessage from './components/RegistrationFailure';
 import {
   backupRegistrationFormBegin,
   clearRegistrationBackendError,
   registerNewUser,
   setUserPipelineDataLoaded,
 } from './data/actions';
-import {
-  FIELDS, FORBIDDEN_REQUEST, INTERNAL_SERVER_ERROR, TPA_AUTHENTICATION_FAILURE, TPA_SESSION_EXPIRED,
-} from './data/constants';
+import { INTERNAL_SERVER_ERROR } from './data/constants';
 import RegistrationPage from './RegistrationPage';
 import {
-  AUTHN_PROGRESSIVE_PROFILING, COMPLETE_STATE, LOGIN_PAGE, PENDING_STATE, REGISTER_PAGE,
+  AUTHN_PROGRESSIVE_PROFILING, COMPLETE_STATE, PENDING_STATE, REGISTER_PAGE,
 } from '../data/constants';
 
 jest.mock('@edx/frontend-platform/analytics', () => ({
@@ -36,7 +33,6 @@ jest.mock('@edx/frontend-platform/i18n', () => ({
 }));
 
 const IntlRegistrationPage = injectIntl(RegistrationPage);
-const IntlRegistrationFailure = injectIntl(RegistrationFailureMessage);
 const mockStore = configureStore();
 
 jest.mock('react-router-dom', () => {
@@ -95,7 +91,6 @@ describe('RegistrationPage', () => {
     currentProvider: null,
     finishAuthUrl: null,
     providers: [],
-    secondaryProviders: [],
     pipelineUserDetails: null,
     countryCode: null,
   };
@@ -152,14 +147,6 @@ describe('RegistrationPage', () => {
     }
   };
 
-  const ssoProvider = {
-    id: 'oa2-apple-id',
-    name: 'Apple',
-    iconClass: 'apple',
-    iconImage: 'https://openedx.devstack.lms/logo.png',
-    loginUrl: '/auth/login/apple-id/?auth_entry=login&next=/dashboard',
-  };
-
   describe('Test Registration Page', () => {
     mergeConfig({
       SHOW_CONFIGURABLE_EDX_FIELDS: true,
@@ -171,10 +158,6 @@ describe('RegistrationPage', () => {
       email: 'Enter your email',
       password: 'Password criteria has not been met',
       country: 'Select your country or region of residence',
-    };
-
-    const secondaryProviders = {
-      id: 'saml-test', name: 'Test University', loginUrl: '/dummy-auth', registerUrl: '/dummy_auth',
     };
 
     // ******** test registration form submission ********
@@ -213,7 +196,7 @@ describe('RegistrationPage', () => {
         email: 'john.doe@example.com',
         country: 'Pakistan',
         honor_code: true,
-        social_auth_provider: ssoProvider.name,
+        social_auth_provider: 'Apple',
         totalRegistrationTime: 0,
       };
 
@@ -223,7 +206,7 @@ describe('RegistrationPage', () => {
           ...initialState.commonComponents,
           thirdPartyAuthContext: {
             ...initialState.commonComponents.thirdPartyAuthContext,
-            currentProvider: ssoProvider.name,
+            currentProvider: 'Apple',
           },
         },
       });
@@ -341,81 +324,6 @@ describe('RegistrationPage', () => {
       expect(store.dispatch).toHaveBeenCalledWith(clearRegistrationBackendError('email'));
     });
 
-    // ******** test alert messages ********
-
-    it('should match third party auth alert', () => {
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            currentProvider: 'Apple',
-          },
-        },
-      });
-
-      const expectedMessage = `${'You\'ve successfully signed into Apple! We just need a little more information before '
-                              + 'you start learning with '}${ getConfig().SITE_NAME }.`;
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find('#tpa-alert').find('p').text()).toEqual(expectedMessage);
-    });
-
-    it('should match internal server error message', () => {
-      const expectedMessage = 'We couldn\'t create your account.An error has occurred. Try refreshing the page, or check your internet connection.';
-      props = {
-        errorCode: INTERNAL_SERVER_ERROR,
-        failureCount: 0,
-      };
-
-      const registrationPage = mount(reduxWrapper(<IntlRegistrationFailure {...props} />));
-      expect(registrationPage.find('div.alert-heading').length).toEqual(1);
-      expect(registrationPage.find('div.alert').first().text()).toEqual(expectedMessage);
-    });
-
-    it('should match registration api rate limit error message', () => {
-      const expectedMessage = 'We couldn\'t create your account.Too many failed registration attempts. Try again later.';
-      props = {
-        errorCode: FORBIDDEN_REQUEST,
-        failureCount: 0,
-      };
-
-      const registrationPage = mount(reduxWrapper(<IntlRegistrationFailure {...props} />));
-      expect(registrationPage.find('div.alert-heading').length).toEqual(1);
-      expect(registrationPage.find('div.alert').first().text()).toEqual(expectedMessage);
-    });
-
-    it('should match tpa session expired error message', () => {
-      const expectedMessage = 'We couldn\'t create your account.Registration using Google has timed out.';
-      props = {
-        context: {
-          provider: 'Google',
-        },
-        errorCode: TPA_SESSION_EXPIRED,
-        failureCount: 0,
-      };
-
-      const registrationPage = mount(reduxWrapper(<IntlRegistrationFailure {...props} />));
-      expect(registrationPage.find('div.alert-heading').length).toEqual(1);
-      expect(registrationPage.find('div.alert').first().text()).toEqual(expectedMessage);
-    });
-
-    it('should match tpa authentication failed error message', () => {
-      const expectedMessageSubstring = 'We are sorry, you are not authorized to access';
-      props = {
-        context: {
-          provider: 'Google',
-        },
-        errorCode: TPA_AUTHENTICATION_FAILURE,
-        failureCount: 0,
-      };
-
-      const registrationPage = mount(reduxWrapper(<IntlRegistrationFailure {...props} />));
-      expect(registrationPage.find('div.alert-heading').length).toEqual(1);
-      expect(registrationPage.find('div.alert').first().text()).toContain(expectedMessageSubstring);
-    });
-
     // ******** test form buttons and fields ********
 
     it('should match default button state', () => {
@@ -451,80 +359,12 @@ describe('RegistrationPage', () => {
       });
     });
 
-    it('should show single sign on provider button', () => {
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            providers: [ssoProvider],
-          },
-        },
-      });
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find(`button#${ssoProvider.id}`).length).toEqual(1);
-    });
-
-    it('should display institution register button', () => {
-      mergeConfig({
-        DISABLE_ENTERPRISE_LOGIN: 'true',
-      });
-
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            secondaryProviders: [secondaryProviders],
-          },
-        },
-      });
-      delete window.location;
-      window.location = { href: getConfig().BASE_URL };
-
-      const root = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(root.text().includes('Institution/campus credentials')).toBe(true);
-
-      mergeConfig({
-        DISABLE_ENTERPRISE_LOGIN: '',
-      });
-    });
-
-    it('should display InstitutionLogistration if insitutionLogin prop is true', () => {
-      props = {
-        ...props,
-        institutionLogin: true,
-      };
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find('.institutions__heading').text()).toEqual('Register with institution/campus credentials');
-    });
-
     it('should show button label based on cta query params value', () => {
       const buttonLabel = 'Register';
       delete window.location;
       window.location = { href: getConfig().BASE_URL, search: `?cta=${buttonLabel}` };
       const registrationPage = mount(reduxWrapper(<IntlRegistrationPage {...props} />));
       expect(registrationPage.find('button[type="submit"] span').first().text()).toEqual(buttonLabel);
-    });
-
-    it('should not display password field when current provider is present', () => {
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            currentProvider: ssoProvider.name,
-          },
-        },
-      });
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find('input#password').length).toEqual(0);
     });
 
     it('should check user retention cookie', () => {
@@ -558,57 +398,6 @@ describe('RegistrationPage', () => {
       window.location = { href: getConfig().BASE_URL };
       renderer.create(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
       expect(window.location.href).toBe(dashboardURL);
-    });
-
-    it('should redirect to social auth provider url on SSO button click', () => {
-      const registerUrl = '/auth/login/apple-id/?auth_entry=register&next=/dashboard';
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            providers: [{
-              ...ssoProvider,
-              registerUrl,
-            }],
-          },
-        },
-      });
-
-      delete window.location;
-      window.location = { href: getConfig().BASE_URL };
-
-      const loginPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-
-      loginPage.find('button#oa2-apple-id').simulate('click');
-      expect(window.location.href).toBe(getConfig().LMS_BASE_URL + registerUrl);
-    });
-
-    it('should redirect to finishAuthUrl upon successful registration via SSO', () => {
-      const authCompleteUrl = '/auth/complete/google-oauth2/';
-      store = mockStore({
-        ...initialState,
-        register: {
-          ...initialState.register,
-          registrationResult: {
-            success: true,
-          },
-        },
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            finishAuthUrl: authCompleteUrl,
-          },
-        },
-      });
-
-      delete window.location;
-      window.location = { href: getConfig().BASE_URL };
-
-      renderer.create(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(window.location.href).toBe(getConfig().LMS_BASE_URL + authCompleteUrl);
     });
 
     it('should redirect to dashboard if features flags are configured but no optional fields are configured', () => {
@@ -670,112 +459,6 @@ describe('RegistrationPage', () => {
       ));
       progressiveProfilingPage.update();
       expect(mockNavigate).toHaveBeenCalledWith(AUTHN_PROGRESSIVE_PROFILING);
-    });
-
-    // ******** test hinted third party auth ********
-
-    it('should render tpa button for tpa_hint id matching one of the primary providers', () => {
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            providers: [ssoProvider],
-          },
-          thirdPartyAuthApiStatus: COMPLETE_STATE,
-        },
-      });
-
-      delete window.location;
-      window.location = { href: getConfig().BASE_URL.concat(LOGIN_PAGE), search: `?next=/dashboard&tpa_hint=${ssoProvider.id}` };
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find(`button#${ssoProvider.id}`).find('span').text()).toEqual(ssoProvider.name);
-      expect(registrationPage.find(`button#${ssoProvider.id}`).hasClass(`btn-tpa btn-${ssoProvider.id}`)).toEqual(true);
-    });
-
-    it('should display skeleton if tpa_hint is true and thirdPartyAuthContext is pending', () => {
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthApiStatus: PENDING_STATE,
-        },
-      });
-
-      delete window.location;
-      window.location = {
-        href: getConfig().BASE_URL.concat(LOGIN_PAGE),
-        search: `?next=/dashboard&tpa_hint=${ssoProvider.id}`,
-      };
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find('.react-loading-skeleton').exists()).toBeTruthy();
-    });
-
-    it('should render icon if icon classes are missing in providers', () => {
-      ssoProvider.iconClass = null;
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            providers: [ssoProvider],
-          },
-          thirdPartyAuthApiStatus: COMPLETE_STATE,
-        },
-      });
-
-      delete window.location;
-      window.location = { href: getConfig().BASE_URL.concat(REGISTER_PAGE), search: `?next=/dashboard&tpa_hint=${ssoProvider.id}` };
-      ssoProvider.iconImage = null;
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find(`button#${ssoProvider.id}`).find('div').find('span').hasClass('pgn__icon')).toEqual(true);
-    });
-
-    it('should render tpa button for tpa_hint id matching one of the secondary providers', () => {
-      secondaryProviders.skipHintedLogin = true;
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            secondaryProviders: [secondaryProviders],
-          },
-          thirdPartyAuthApiStatus: COMPLETE_STATE,
-        },
-      });
-
-      delete window.location;
-      window.location = { href: getConfig().BASE_URL.concat(REGISTER_PAGE), search: `?next=/dashboard&tpa_hint=${secondaryProviders.id}` };
-
-      mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(window.location.href).toEqual(getConfig().LMS_BASE_URL + secondaryProviders.registerUrl);
-    });
-
-    it('should render regular tpa button for invalid tpa_hint value', () => {
-      const expectedMessage = `${ssoProvider.name}`;
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            providers: [ssoProvider],
-          },
-          thirdPartyAuthApiStatus: COMPLETE_STATE,
-        },
-      });
-
-      delete window.location;
-      window.location = { href: getConfig().BASE_URL.concat(LOGIN_PAGE), search: '?next=/dashboard&tpa_hint=invalid' };
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find(`button#${ssoProvider.id}`).find('span#provider-name').text()).toEqual(expectedMessage);
     });
 
     // ******** miscellaneous tests ********
@@ -998,7 +681,6 @@ describe('RegistrationPage', () => {
           thirdPartyAuthApiStatus: COMPLETE_STATE,
           thirdPartyAuthContext: {
             ...initialState.commonComponents.thirdPartyAuthContext,
-            currentProvider: ssoProvider.name,
             pipelineUserDetails: {
               name: 'John Doe',
               username: 'john_doe',
@@ -1046,7 +728,7 @@ describe('RegistrationPage', () => {
           thirdPartyAuthApiStatus: COMPLETE_STATE,
           thirdPartyAuthContext: {
             ...initialState.commonComponents.thirdPartyAuthContext,
-            currentProvider: ssoProvider.name,
+            currentProvider: 'Apple',
             pipelineUserDetails: {
               name: 'John Doe',
               username: 'john_doe',
@@ -1067,164 +749,6 @@ describe('RegistrationPage', () => {
         social_auth_provider: 'Apple',
         totalRegistrationTime: 0,
       }));
-    });
-
-    it('should display errorMessage if third party authentication fails', () => {
-      jest.spyOn(global.Date, 'now').mockImplementation(() => 0);
-      getLocale.mockImplementation(() => ('en-us'));
-
-      store = mockStore({
-        ...initialState,
-        register: {
-          ...initialState.register,
-          backendCountryCode: 'PK',
-          userPipelineDataLoaded: false,
-        },
-        commonComponents: {
-          ...initialState.commonComponents,
-          thirdPartyAuthApiStatus: COMPLETE_STATE,
-          thirdPartyAuthContext: {
-            ...initialState.commonComponents.thirdPartyAuthContext,
-            currentProvider: null,
-            pipelineUserDetails: {},
-            errorMessage: 'An error occurred',
-          },
-        },
-      });
-
-      store.dispatch = jest.fn(store.dispatch);
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find('div.alert-heading').length).toEqual(1);
-      expect(registrationPage.find('div.alert').first().text()).toContain('An error occurred');
-    });
-  });
-
-  describe('Test Configurable Fields', () => {
-    mergeConfig({
-      ENABLE_DYNAMIC_REGISTRATION_FIELDS: true,
-      SHOW_CONFIGURABLE_EDX_FIELDS: true,
-    });
-
-    it('should render fields returned by backend', () => {
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          fieldDescriptions: {
-            profession: { name: 'profession', type: 'text', label: 'Profession' },
-            terms_of_service: {
-              name: FIELDS.TERMS_OF_SERVICE,
-              error_message: 'You must agree to the Terms and Service agreement of our site',
-            },
-          },
-        },
-      });
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      expect(registrationPage.find('#profession').exists()).toBeTruthy();
-      expect(registrationPage.find('#tos').exists()).toBeTruthy();
-    });
-
-    it('should submit form with fields returned by backend in payload', () => {
-      getLocale.mockImplementation(() => ('en-us'));
-      jest.spyOn(global.Date, 'now').mockImplementation(() => 0);
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          fieldDescriptions: {
-            profession: { name: 'profession', type: 'text', label: 'Profession' },
-          },
-          extendedProfile: ['profession'],
-        },
-      });
-
-      const payload = {
-        name: 'John Doe',
-        username: 'john_doe',
-        email: 'john.doe@example.com',
-        password: 'password1',
-        country: 'Pakistan',
-        honor_code: true,
-        profession: 'Engineer',
-        totalRegistrationTime: 0,
-      };
-
-      store.dispatch = jest.fn(store.dispatch);
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-
-      populateRequiredFields(registrationPage, payload);
-      registrationPage.find('input#profession').simulate('change', { target: { value: 'Engineer', name: 'profession' } });
-      registrationPage.find('button.btn-brand').simulate('click');
-      expect(store.dispatch).toHaveBeenCalledWith(registerNewUser({ ...payload, country: 'PK' }));
-    });
-
-    it('should show error messages for required fields on empty form submission', () => {
-      const professionError = 'Enter your profession';
-      const countryError = 'Select your country or region of residence';
-      const confirmEmailError = 'Enter your email';
-
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          fieldDescriptions: {
-            profession: {
-              name: 'profession', type: 'text', label: 'Profession', error_message: professionError,
-            },
-            confirm_email: {
-              name: 'confirm_email', type: 'text', label: 'Confirm Email', error_message: confirmEmailError,
-            },
-            country: { name: 'country' },
-          },
-        },
-      });
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      registrationPage.find('button.btn-brand').simulate('click');
-
-      expect(registrationPage.find('#profession-error').last().text()).toEqual(professionError);
-      expect(registrationPage.find('div[feedback-for="country"]').text()).toEqual(countryError);
-      expect(registrationPage.find('#confirm_email-error').last().text()).toEqual(confirmEmailError);
-    });
-
-    it('should show error if email and confirm email fields do not match', () => {
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          fieldDescriptions: {
-            confirm_email: {
-              name: 'confirm_email', type: 'text', label: 'Confirm Email',
-            },
-          },
-        },
-      });
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      registrationPage.find('input#email').simulate('change', { target: { value: 'test1@gmail.com', name: 'email' } });
-      registrationPage.find('input#confirm_email').simulate('blur', { target: { value: 'test2@gmail.com', name: 'confirm_email' } });
-      expect(registrationPage.find('div#confirm_email-error').text()).toEqual('The email addresses do not match.');
-    });
-
-    it('should run validations for configurable focused field on form submission', () => {
-      const professionError = 'Enter your profession';
-      store = mockStore({
-        ...initialState,
-        commonComponents: {
-          ...initialState.commonComponents,
-          fieldDescriptions: {
-            profession: {
-              name: 'profession', type: 'text', label: 'Profession', error_message: professionError,
-            },
-          },
-        },
-      });
-
-      const registrationPage = mount(routerWrapper(reduxWrapper(<IntlRegistrationPage {...props} />)));
-      registrationPage.find('input#profession').simulate('focus', { target: { value: '', name: 'profession' } });
-      registrationPage.find('button.btn-brand').simulate('click');
-
-      expect(registrationPage.find('#profession-error').last().text()).toEqual(professionError);
     });
   });
 });
